+++ src/notification-center/NotificationArchive/index.js
@@ -1,0 +1,1 @@
+export { default as NotificationArchive } from './NotificationArchive.svelte';