+++ src/notification-center/NotificationCenter/index.js
@@ -1,0 +1,2 @@
+export { default as NotificationCenter } from './NotificationCenter.svelte';
+export * from './NotificationCenter.js';